feat: editable, per-location quantities in the All Items detail view
Switch LocationItem.quantity to Double so items can be partially
consumed (e.g. half a bag of flour), and replace the read-only detail
pushed from the All Items tab with AllItemsProductDetailView, which
lists every location holding the product with its own quantity
stepper (labeling rows by location only when there's more than one).

+++ Foodpoint/Models/Location.swift
@@ -18,5 +18,5 @@ struct Location: Identifiable {
 struct LocationItem: Identifiable {
     let id: String
     var product: FoodProduct
-    var quantity: Int
+    var quantity: Double
 }
